+++ front/src/widgets/Progressbar/ui/Progressbar/Progressbar.tsx
@@ -1,6 +1,15 @@
 import cls from './Progressbar.module.scss';
-import { Progressstep } from '../Progressstep/Progressstep';
-import { ProgressSteps } from '../../model/type/progressbar';
+import { ProgressbarStep } from '../ProgressbarStep/ProgressbarStep';
+
+export const ProgressSteps: { id: string, title: string }[] = [
+  { 'id': '1', 'title': 'Новая' },
+  { 'id': '2', 'title': 'Обрабатывается' },
+  { 'id': '3', 'title': 'На согласовании у заказчика' },
+  { 'id': '4', 'title': 'Стоимость работ' },
+  { 'id': '5', 'title': 'В работе' },
+  { 'id': '6', 'title': 'На согласовании у заказчика' },
+  { 'id': '7', 'title': 'Сделано' },
+];
 
 interface ProgressbarProps {
   className?: string;
@@ -16,7 +25,7 @@ export const Progressbar: React.FC<ProgressbarProps> = (props) => {
     <div className={cls.progressbar}>
       <ul className={cls.list}>
         {ProgressSteps.map((item) =>
-          <Progressstep step={step} id={item.id} title={item.title} />
+          <ProgressbarStep step={step} id={item.id} title={item.title} key={item.id} />
         )}
       </ul>
     </div>
