+++ front/src/widgets/Progressbar/ui/ProgressbarStep/ProgressbarStep.module.scss
